+++ internal/metadata/metadata.go
@@ -1,8 +1,30 @@
 package metadata
 
-import "strings"
+import (
+	"strings"
+
+	"github.com/superblocksteam/agent/pkg/utils"
+)
 
 func GetTagsMap(tagsString string) map[string][]string {
+	return parseTagsString(tagsString)
+}
+
+func GetTagsSetMap(tagsString string) map[string]*utils.Set[string] {
+	tagsMap := parseTagsString(tagsString)
+	if len(tagsMap) == 0 {
+		return make(map[string]*utils.Set[string])
+	}
+
+	tagsSetMap := make(map[string]*utils.Set[string], len(tagsMap))
+	for key, values := range tagsMap {
+		tagsSetMap[key] = utils.NewSet(values...)
+	}
+
+	return tagsSetMap
+}
+
+func parseTagsString(tagsString string) map[string][]string {
 	if len(tagsString) == 0 {
 		return make(map[string][]string)
 	}